moshin_backend/api/authentication/serializers/response.py
from rest_framework import serializers


class LoginResponseSerializer(serializers.Serializer):
    access = serializers.CharField()
    refresh = serializers.CharField()
    user_id = serializers.UUIDField()
